+++ farmtest/src/main/java/com/kosta/farm/repository/OrdersRepository.java
@@ -3,7 +3,7 @@ package com.kosta.farm.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kosta.farm.entity.Orders;
+import com.kosta.farm.entity.Payment;
 
-public interface OrdersRepository extends JpaRepository<Orders, Integer> {
-
+public interface OrdersRepository extends JpaRepository<Orders, Long> {
 }
