+++ farmtest/src/main/java/com/kosta/farm/repository/ProductFileRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kosta.farm.entity.ProductFile;
 
-public interface ProductFileRepository extends JpaRepository<ProductFile, Integer> {
+public interface ProductFileRepository extends JpaRepository<ProductFile, Long> {
 
 }
