+++ farmtest/src/main/java/com/kosta/farm/repository/QuotationRepositrory.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kosta.farm.entity.Quotation;
 
-public interface QuotationRepositrory extends JpaRepository<Quotation, Integer> {
+public interface QuotationRepositrory extends JpaRepository<Quotation, Long> {
 
 	
 }
